% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_tgz.R
\name{artifact_folder_paths}
\alias{artifact_folder_paths}
\title{obtain all package artifact folder paths}
\usage{
artifact_folder_paths(
  version = "3.14",
  type = "bioc",
  cache = BiocFileCache::BiocFileCache(),
  url = NULL,
  destbase = "report"
)
}
\arguments{
\item{version}{character(1) defaults to "3.14"}

\item{type}{character(1) defaults to 'bioc' which implies 'software'; see Note.}

\item{cache}{instance of `BiocFileCache::BiocFileCache()`}

\item{url}{passed to `get_report_tgz_cacheid`}

\item{destbase}{passed to `path_to_untarred_artifact_folders`, defaults to "report"}
}
\description{
obtain all package artifact folder paths
}
\examples{
cururl = demo_url()
artifact_folder_paths(url=cururl, destbase="test_report")
}
