% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_artifacts.R
\name{as.data.frame.ArtifSet}
\alias{as.data.frame.ArtifSet}
\title{basic artifact data collected to tibble}
\usage{
\method{as.data.frame}{ArtifSet}(x, row.names = NULL, optional = FALSE, ...)
}
\description{
basic artifact data collected to tibble
}
