% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artif_store.R
\name{available_artifs}
\alias{available_artifs}
\title{tabulate available artifacts in a cache}
\usage{
available_artifs(ca = BiocFileCache::BiocFileCache(), qstring = "report.tgz")
}
\arguments{
\item{ca}{cache reference}

\item{qstring}{character(1) search key}
}
\description{
tabulate available artifacts in a cache
}
