% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_package_data.R
\docType{class}
\name{BBS_package_data-class}
\alias{BBS_package_data-class}
\title{encapsulate BBS information from multiple platforms for a package}
\description{
encapsulate BBS information from multiple platforms for a package
}
