% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_package_data.R
\docType{class}
\name{BBS_pkg_data_for_host-class}
\alias{BBS_pkg_data_for_host-class}
\title{encapsulate data on BBS state for a single host}
\description{
encapsulate data on BBS state for a single host
}
