% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_events.R
\name{browse_event_collection}
\alias{browse_event_collection}
\title{app to get details on events in packages}
\usage{
browse_event_collection(ec, af)
}
\arguments{
\item{ec}{event_collection S3 instance}

\item{af}{ArtifSet S4 instance}
}
\description{
app to get details on events in packages
}
