% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_tgz.R
\name{build_report_tgz_url}
\alias{build_report_tgz_url}
\title{make url for a checkResults instance}
\usage{
build_report_tgz_url(version, type)
}
\arguments{
\item{version}{character(1)}

\item{type}{character(1)}
}
\description{
make url for a checkResults instance
}
