% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_tgz.R
\name{demo_url2}
\alias{demo_url2}
\title{A demonstration URL with trimmed report.tgz in a local file, more recent than that for `demo_url`}
\usage{
demo_url2()
}
\description{
A demonstration URL with trimmed report.tgz in a local file, more recent than that for `demo_url`
}
