% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_freqs.R
\name{event_freqs}
\alias{event_freqs}
\title{provide a concise table of events by host}
\usage{
event_freqs(
  af,
  hostmap = c(nebbiolo2 = "linux", machv2 = "mac", tokay2 = "win")
)
}
\arguments{
\item{af}{ArtifSet instance}

\item{hostmap}{character() vector with values c("linux", "win", "mac") and names the names of hosts in use}
}
\description{
provide a concise table of events by host
}
\examples{
af = make_demo_ArtifSet()
event_freqs(af)  # a package can generate events in multiple phases, so counts are not of 'packages'
}
