% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_cache.R
\name{examine_cache}
\alias{examine_cache}
\title{look at a cache for potential BBS build artifact tarballs}
\usage{
examine_cache(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{BiocFileCache instance, defaults to BiocFileCache::BiocFileCache()}
}
\description{
look at a cache for potential BBS build artifact tarballs
}
