% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_artifacts.R
\name{get_events}
\alias{get_events}
\title{get events from an artifact set using rcmdcheck::parse_check}
\usage{
get_events(apathset, type)
}
\arguments{
\item{apathset}{instance of artifact_paths}

\item{type}{character(1) one of "error", "warning", "note"}
}
\description{
get events from an artifact set using rcmdcheck::parse_check
}
\examples{
setup_demo_artifacts()
eg = select_artifacts(paste(tempdir(), "tokay2", sep="/"),
     c("IRanges", "S4V*", "a4\\\\."))
get_events(eg)
}
