% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_tgz.R
\name{get_report_tgz_cacheid}
\alias{get_report_tgz_cacheid}
\title{get reporting artifacts for a Bioconductor collection (e.g., software, experiment, workflow, ...)}
\usage{
get_report_tgz_cacheid(
  version = "3.14",
  type = "bioc",
  cache = BiocFileCache::BiocFileCache()
)
}
\arguments{
\item{version}{character(1) defaults to "3.14"}

\item{type}{character(1) defaults to 'bioc' which implies 'software'; see Note.}

\item{cache}{instance of `BiocFileCache::BiocFileCache()`}
}
\value{
A gzipped tarball is downloaded, copied to a cache, and the cache reference is returned.
}
\description{
get reporting artifacts for a Bioconductor collection (e.g., software, experiment, workflow, ...)
}
\note{
Use bbsBuildArtifacts:::valid_types() to see valid values for `type`.
}
