% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_tgz.R
\name{get_report_tgz_cacheid2}
\alias{get_report_tgz_cacheid2}
\title{try with date}
\usage{
get_report_tgz_cacheid2(
  version = "3.14",
  type = "bioc",
  cache = BiocFileCache::BiocFileCache(),
  url = NULL,
  date = Sys.Date()
)
}
\description{
try with date
}
