% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_package_data.R
\name{host_data_by_phase}
\alias{host_data_by_phase}
\title{acquire data on a package's BBS state}
\usage{
host_data_by_phase(bbspd, host, phase)
}
\arguments{
\item{bbspd}{instance of BBS_package_data}

\item{host}{character(1) valid host name for BBS}

\item{phase}{character(1) valid phase}
}
\description{
acquire data on a package's BBS state
}
\examples{
af = make_demo_ArtifSet()
ho = slot(af, "hostnames")
ho
pd1 = make_BBS_package_data(af, "zinbwave")
hd = host_data_by_phase( pd1, ho[1], "buildsrc")
head(hd)
hd = host_data_by_phase( pd1, ho[1], "checksrc")
head(hd)
pd2 = make_BBS_package_data(af, "zlibbioc")
hd = host_data_by_phase( pd2, ho[1], "install")
tail(hd)
hd = host_data_by_phase( pd2, ho[1], "checksrc")
head(hd)
}
