% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_artifacts.R
\name{hostnames_by_release}
\alias{hostnames_by_release}
\title{vector of hostnames for build nodes}
\usage{
hostnames_by_release(release = "3.16")
}
\arguments{
\item{release}{character(1) defaults to '3.16'}
}
\description{
vector of hostnames for build nodes
}
