% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idiosync_status.R
\name{idiosync_status}
\alias{idiosync_status}
\title{enumerate packages that fail on only a subset of platforms}
\usage{
idiosync_status(af, phase.in = "checksrc", state = "ERROR")
}
\arguments{
\item{af}{ArtifSet instance}

\item{phase.in}{character(1) e.g., "checksrc", "buildbin", "install"}

\item{state}{character(1) e.g., "ERROR", "WARNINGS"}
}
\description{
enumerate packages that fail on only a subset of platforms
}
\examples{
af = make_demo_ArtifSet()
idiosync_status(af)
}
