% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maint4pkg.R
\name{maint4pkg}
\alias{maint4pkg}
\title{given a package name, version, and biocPkgList instance, obtain the email for the maintainer}
\usage{
maint4pkg(pkg, version, bpl)
}
\arguments{
\item{pkg}{character(1)}

\item{version}{character(1)}

\item{bpl}{biocPkgList instance}
}
\description{
given a package name, version, and biocPkgList instance, obtain the email for the maintainer
}
\examples{
bpl = BiocPkgTools::biocPkgList(version="3.17")
maint4pkg("parody", "3.17", bpl)
}
