% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_artifacts.R
\name{make_demo_ArtifSet}
\alias{make_demo_ArtifSet}
\title{simplify production of the test ArtifSet}
\usage{
make_demo_ArtifSet(
  cache = BiocFileCache::BiocFileCache(),
  preclean = TRUE,
  demostring = "demo.tgz",
  url = paste0("file://", demo_path())
)
}
\arguments{
\item{cache}{BiocFileCache instance to use for checking for available image and retrieving if possible}

\item{preclean}{logical(1) defaults to TRUE, in which case we search `cache` for demostring in cache and remove}

\item{demostring}{character(1) token to query BiocFileCache for in preclean step}

\item{url}{character(1) url for local tgz}
}
\description{
simplify production of the test ArtifSet
}
\examples{
make_demo_ArtifSet()
}
