% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_status.R
\name{new_events}
\alias{new_events}
\title{list packages that have a new event comparing af2 (later) to af1 (earlier)}
\usage{
new_events(af1, af2, event = "ERROR", phase = "checksrc", host = "nebbiolo2")
}
\arguments{
\item{af1}{instance of ArtifSet, earlier}

\item{af2}{instance of ArtifSet, later}

\item{event}{character(1) one of "ERROR", "WARNINGS", "OK", "skipped", "TIMEOUT"}

\item{phase}{character(1) one of "install", "buildsrc", "checksrc", "buildbin"}

\item{host}{character(1) relevant hostname}
}
\description{
list packages that have a new event comparing af2 (later) to af1 (earlier)
}
\examples{
\dontrun{    # await new demo
af2 = make_demo_ArtifSet(url = demo_url2(), 
       demostring = "test_report_3.14_bioc_20220105", destbase = "test_report_0105")
af1 = make_demo_ArtifSet()
new_events(af1, af2)
}
}
