% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_states.R
\name{packages_with_events}
\alias{packages_with_events}
\title{get info on packages with non-ok status}
\usage{
packages_with_events(aset)
}
\arguments{
\item{aset}{ArtifSet instance}
}
\value{
a list with one element per phase in which some package produced non-OK state
}
\description{
get info on packages with non-ok status
}
\examples{
z = make_demo_ArtifSet()
packages_with_events(z)
}
