% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_states.R
\name{packnames_with_events}
\alias{packnames_with_events}
\title{filter packages by host, phase, state}
\usage{
packnames_with_events(
  af,
  host = "nebbiolo2",
  phase = "checksrc",
  event_class = "ERROR"
)
}
\arguments{
\item{af}{ArtifSet instance}

\item{host}{character(1) name of build host}

\item{phase}{character(1) one of 'install', 'buildsrc', 'checksrc', 'buildbin'}

\item{event_class}{character(1) one of 'ERROR', 'WARNINGS', 'TIMEOUT', 'skipped', 'wontinstall'}
}
\value{
character vector or NULL
}
\description{
filter packages by host, phase, state
}
\examples{
af = make_demo_ArtifSet()
packnames_with_events(af) # uses a large BUILD_STATUS_DB.txt in demo ArtifSet
}
