% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_tgz.R
\name{path_to_untarred_artifact_folders}
\alias{path_to_untarred_artifact_folders}
\title{untar the artifact archive and return the path}
\usage{
path_to_untarred_artifact_folders(
  version = "3.16",
  type = "bioc",
  cache = BiocFileCache::BiocFileCache(),
  destination = tempdir(),
  verbose = TRUE,
  url = NULL,
  destbase = "/report"
)
}
\arguments{
\item{version}{character(1) defaults to "3.16"}

\item{type}{character(1) defaults to 'bioc' which implies 'software'; see Note.}

\item{cache}{instance of `BiocFileCache::BiocFileCache()`}

\item{destination}{character(1) path to folder to use, defaults to `tempdir()`}

\item{verbose}{logical(1) will indicate start and end of untar() process}

\item{url}{passed to `get_report_tgz_cacheid`}

\item{destbase}{character(1) name of folder that includes all artifacts, defaults to "report"}
}
\value{
character(1) path to folder beneath which all artifacts are found
}
\description{
untar the artifact archive and return the path
}
\note{
Typically some additional files exist in addition to the package-related folders
at the returned path.
`:::non_package_pattern()` can be used to find these.
Use bbsBuildArtifacts:::valid_types() to see valid values for `type`.
}
\examples{
cururl = demo_url()
td = tempdir()
path_to_untarred_artifact_folders(url=cururl, destination=td, destbase="test_report")
}
