% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_artifacts.R
\name{paths,ArtifSet-method}
\alias{paths,ArtifSet-method}
\title{helper subsetter}
\usage{
\S4method{paths}{ArtifSet}(x)
}
\arguments{
\item{x}{instance of ArtifSet}
}
\description{
helper subsetter
}
