% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_status.R
\name{pkgs_with_event}
\alias{pkgs_with_event}
\title{list packages presenting with status on BBS}
\usage{
pkgs_with_event(af, event = "ERROR", phase = "install", host = "nebbiolo2")
}
\arguments{
\item{af}{instance of ArtifSet}

\item{event}{character(1) one of "ERROR", "WARNINGS", "OK", "skipped", "TIMEOUT"}

\item{phase}{character(1) one of "install", "buildsrc", "checksrc", "buildbin"}

\item{host}{character(1) relevant hostname}
}
\description{
list packages presenting with status on BBS
}
\examples{
af = make_demo_ArtifSet()
pkgs_with_event(af)
}
