% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.artifset_scan.R
\name{plot_scan_chk}
\alias{plot_scan_chk}
\title{plot counts of checksrc OK for a scan}
\usage{
plot_scan_chk(x)
}
\arguments{
\item{x}{instance of artifset_scan}
}
\description{
plot counts of checksrc OK for a scan
}
\examples{
data(scan_demo_3.17)
plot_scan_chk(scan_demo_3.17) + ggplot2::ylim(1600,1720)
}
