% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.artifset_scan.R
\name{plot_scan_inst}
\alias{plot_scan_inst}
\title{plot counts of install OK for a scan}
\usage{
plot_scan_inst(x)
}
\arguments{
\item{x}{instance of artifset_scan}
}
\description{
plot counts of install OK for a scan
}
\examples{
data(scan_demo_3.17)
plot_scan_inst(scan_demo_3.17) + ggplot2::ylim(2120,2250)
}
