% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_tgz.R
\name{print.artifact_folder_paths}
\alias{print.artifact_folder_paths}
\title{show artifact paths nicely}
\usage{
\method{print}{artifact_folder_paths}(x, ...)
}
\arguments{
\item{x}{instance of `artifact_folder_paths`}

\item{\dots}{not used}
}
\description{
show artifact paths nicely
}
