% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_artifacts.R
\name{print.artifact_paths}
\alias{print.artifact_paths}
\title{simplify display of artifact set}
\usage{
\method{print}{artifact_paths}(x, ...)
}
\arguments{
\item{x}{instance of artifact_paths}

\item{\dots}{not used}
}
\description{
simplify display of artifact set
}
