% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_report.R
\name{print.bbs_time_report}
\alias{print.bbs_time_report}
\title{print method for time report, focusing on large times}
\usage{
\method{print}{bbs_time_report}(x, ...)
}
\arguments{
\item{x}{instance of bbs_time_report}

\item{\dots}{not used}
}
\description{
print method for time report, focusing on large times
}
