% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_events.R
\name{print.event_collection}
\alias{print.event_collection}
\title{short report of an event collection}
\usage{
\method{print}{event_collection}(x, ...)
}
\arguments{
\item{x}{event_collection instance}

\item{\dots}{not used}
}
\description{
short report of an event collection
}
