% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_tgz.R
\name{print.pkg_by_host_data}
\alias{print.pkg_by_host_data}
\title{simplify presentation of pkg data}
\usage{
\method{print}{pkg_by_host_data}(x, ...)
}
\arguments{
\item{x}{instance of pkg_by_host_data}

\item{\dots}{not used}
}
\description{
simplify presentation of pkg data
}
