% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_r_meta.R
\name{pull_R_meta}
\alias{pull_R_meta}
\title{acquire information on build platform from BBS NodeInfo.html file}
\usage{
pull_R_meta(htmlfile)
}
\arguments{
\item{htmlfile}{character(1) path to NodeInfo.html for a given platform}
}
\description{
acquire information on build platform from BBS NodeInfo.html file
}
\examples{
demaf = make_demo_ArtifSet()
ho = slot(demaf, "hostnames")
ho2_info = grep(ho[2], slot(demaf, "extra_paths"), value=TRUE)
pull_R_meta(grep("NodeInfo", ho2_info, value=TRUE))
}
