% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_614}
\alias{sc_614}
\title{snapshot (6/14/2021) of 3.13 'core' packages processed by 'tabulate_bbs_metrics'}
\format{
data.frame
}
\usage{
sc_614
}
\description{
snapshot (6/14/2021) of 3.13 'core' packages processed by 'tabulate_bbs_metrics'
}
\keyword{datasets}
