% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_artifacts.R
\name{select_artifacts}
\alias{select_artifacts}
\title{collect the paths to a set of artifacts}
\usage{
select_artifacts(hostpath, packs_re)
}
\arguments{
\item{hostpath}{character(1) e.g., path to a folder named 'tokay1', which holds the build artifacts
for this builder}

\item{packs_re}{character() a vector of regular expressions defining packages of interest}
}
\description{
collect the paths to a set of artifacts
}
\note{
Could be improved to work on a comprehensive tarball.
}
\examples{
setup_demo_artifacts()
eg = select_artifacts(paste(tempdir(), "tokay2", sep="/"),
    c("IRanges", "S4V*", "a4\\\\."))
eg
}
