% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_artifacts.R
\name{setup_artifacts}
\alias{setup_artifacts}
\title{Create ArtifSet instance}
\usage{
setup_artifacts(
  type = "bioc",
  version = "3.14",
  date,
  hostnames = hostnames_by_release(version),
  cache = BiocFileCache::BiocFileCache(),
  destination = tempfile(),
  verbose = TRUE,
  extracted = NULL,
  url = NULL,
  destbase = "report"
)
}
\arguments{
\item{type}{character(1) defaults to 'bioc' which implies 'software'; see Note.}

\item{version}{character(1) defaults to "3.14"}

\item{date}{character(1) "yyyy-mm-dd", not obligatory but can be used to retrieve earlier cache entry}

\item{hostnames}{character() vector of host names for which build artifacts are available}

\item{cache}{instance of `BiocFileCache::BiocFileCache()`}

\item{destination}{character(1) path to folder to use for artifacts}

\item{verbose}{logical(1) if TRUE (default) will provide message about tar activity}

\item{extracted}{defaults to NULL, if non-null a character(1) path to folder that holds `report` folder}

\item{url}{character(1) passed to `get_report_tgz_cacheid`}

\item{destbase}{character(1) name of folder holding all artifacts, defaults to 'report'}
}
\description{
Create ArtifSet instance
}
\note{
Use bbsBuildArtifacts:::valid_types() to see valid values for `type`.  The logic
of managing artifacts from multiple dates is very cumbersome and may be unreliable,
but the 'date of tarball production' will be reported for each ArtifSet.
}
\examples{
cururl = demo_url()
z = setup_artifacts(url=cururl, destbase="test_report")
z
}
