% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_artifacts.R
\name{setup_demo_artifacts}
\alias{setup_demo_artifacts}
\title{.Rcheck artifacts are not reliably installed from inst, so we use a zip file}
\usage{
setup_demo_artifacts()
}
\description{
.Rcheck artifacts are not reliably installed from inst, so we use a zip file
}
