% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_artifacts.R
\name{show,ArtifSet-method}
\alias{show,ArtifSet-method}
\title{display salient information about ArtifSet}
\usage{
\S4method{show}{ArtifSet}(object)
}
\description{
display salient information about ArtifSet
}
