% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_package_data.R
\name{show,BBS_package_data-method}
\alias{show,BBS_package_data-method}
\title{simple display of BBS package data}
\usage{
\S4method{show}{BBS_package_data}(object)
}
\arguments{
\item{object}{instance of BBS_package_data}
}
\description{
simple display of BBS package data
}
