% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_package_data.R
\name{show,BBS_pkg_data_for_host-method}
\alias{show,BBS_pkg_data_for_host-method}
\title{simple display of BBS package data for host}
\usage{
\S4method{show}{BBS_pkg_data_for_host}(object)
}
\arguments{
\item{object}{instance of BBS_pkg_data_for_host}
}
\description{
simple display of BBS package data for host
}
