% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_package_data.R
\name{show,BBS_raw_pkg_info-method}
\alias{show,BBS_raw_pkg_info-method}
\title{simple display of BBS raw package info from info.dcf}
\usage{
\S4method{show}{BBS_raw_pkg_info}(object)
}
\arguments{
\item{object}{instance of BBS_raw_pkg_info}
}
\description{
simple display of BBS raw package info from info.dcf
}
