% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simplecore}
\alias{simplecore}
\title{subjective listing of core packages, mostly by maintainer@bioconductor}
\format{
data.frame
}
\usage{
simplecore
}
\description{
subjective listing of core packages, mostly by maintainer@bioconductor
}
\keyword{datasets}
