% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_tgz.R
\name{simplify_artifact_build_dcfs}
\alias{simplify_artifact_build_dcfs}
\title{produce a data.frame from a collection of `package_by_host_data` outputs}
\usage{
simplify_artifact_build_dcfs(x, ...)
}
\arguments{
\item{x}{instance of `artifact_build_dcfs`}
}
\description{
produce a data.frame from a collection of `package_by_host_data` outputs
}
