% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_report.R
\name{summary.bbs_time_report}
\alias{summary.bbs_time_report}
\title{print method for time report, focusing on large times}
\usage{
\method{summary}{bbs_time_report}(object, ...)
}
\arguments{
\item{object}{instance of bbs_time_report}

\item{\dots}{not used}
}
\description{
print method for time report, focusing on large times
}
