% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulate_bbs_metrics}
\alias{tabulate_bbs_metrics}
\title{tabulate key metrics for a set of packages for a host}
\usage{
tabulate_bbs_metrics(
  packre,
  hostpath,
  procs = c("install", "buildsrc", "checksrc")
)
}
\arguments{
\item{packre}{character() vector of regular expressions picking out
packages in a folder of build artifacts.  For example `"a4\\."` can be used 
as an element of `packre` to isolate package `a4`, while `"a4"` would
select both `a4` and `a4Core`.}

\item{hostpath}{character(1) path to folder, named for host where packages were built to generate artifacts,
in which all artifacts reside}

\item{procs}{character() processes to report, from among "install", "buildsrc", "buildbin" (only
relevant for hosts on which binary packages are built), "checksrc"}
}
\description{
tabulate key metrics for a set of packages for a host
}
\note{
Early implementation just tabulates time and checksrc status.
}
\examples{
packre = c("IRanges", "S4V", "a4")
host = "nebbiolo1"
setup_demo_artifacts()
hostpath = paste(tempdir(), host, sep="/")
chk = tabulate_bbs_metrics( packre, hostpath )
chk
}
