% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_states.R
\name{tabulate_states}
\alias{tabulate_states}
\title{(can be memoised) retrieve data.frame from BUILD_STATUS_DB.txt}
\usage{
tabulate_states(aset)
}
\arguments{
\item{aset}{ArtifSet instance}
}
\description{
(can be memoised) retrieve data.frame from BUILD_STATUS_DB.txt
}
\note{
Restricts attention to packages in available paths.
}
