% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_report.R
\name{time_report}
\alias{time_report}
\title{produce a simple list with data on times}
\usage{
time_report(af, num = 10)
}
\arguments{
\item{af}{ArtifSet instance}

\item{num}{number of largest times to report per phase, default to 10}
}
\description{
produce a simple list with data on times
}
\examples{
af = make_demo_ArtifSet()
r = time_report(af)
r
summary(r)
}
