% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlink_artifset.R
\name{unlink_artifset}
\alias{unlink_artifset}
\title{remove all elements of an ArtifSet from disk}
\usage{
unlink_artifset(aset)
}
\arguments{
\item{aset}{ArtifSet instance}
}
\description{
remove all elements of an ArtifSet from disk
}
