% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_tgz.R
\name{artifact_folder_paths}
\alias{artifact_folder_paths}
\title{obtain all package artifact folder paths}
\usage{
artifact_folder_paths(
  version = "3.14",
  type = "bioc",
  cache = BiocFileCache::BiocFileCache()
)
}
\arguments{
\item{version}{character(1) defaults to "3.14"}

\item{type}{character(1) defaults to 'bioc' which implies 'software'; see Note.}

\item{cache}{instance of `BiocFileCache::BiocFileCache()`}
}
\description{
obtain all package artifact folder paths
}
