% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_artifacts.R
\name{as.data.frame.ArtifSet}
\alias{as.data.frame.ArtifSet}
\title{basic artifact data collected to tibble}
\usage{
\method{as.data.frame}{ArtifSet}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{instance of ArtifSet}

\item{row.names}{defaults to NULL}

\item{optional}{defaults to FALSE}

\item{\dots}{unused}
}
\description{
basic artifact data collected to tibble
}
