% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_events.R
\name{browse_events}
\alias{browse_events}
\title{app to get details on events in packages}
\usage{
browse_events(af)
}
\arguments{
\item{af}{ArtifSet S4 instance}
}
\description{
app to get details on events in packages
}
\examples{
af = make_demo_ArtifSet()
browse_events(af)
}
