% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_events.R
\name{browse_events}
\alias{browse_events}
\title{app to get details on events in packages}
\usage{
browse_events(af)
}
\arguments{
\item{af}{ArtifSet S4 instance
element values are the host names used in the construction of html in package build report.}
}
\description{
app to get details on events in packages
}
\examples{
af = make_demo_ArtifSet()
if (interactive()) {
 browse_events(af)
}
}
