% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_events.R
\name{browse_events}
\alias{browse_events}
\title{app to get details on events in packages}
\usage{
browse_events(version = "3.15")
}
\arguments{
\item{version}{character(1) defaults to "3.15"}
}
\description{
app to get details on events in packages
}
\examples{
if (interactive()) {
 browse_events()
}
}
