% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_events.R
\name{collect_events}
\alias{collect_events}
\title{scan a slice of an ArtifSet for events produced in R CMD check of source}
\usage{
collect_events(
  af,
  event_class = "warnings",
  BPPARAM = BiocParallel::bpparam(),
  which_to_use = NULL
)
}
\arguments{
\item{af}{instance of ArtifSet}

\item{event_class}{character(1) "warnings" (default) or "errors"}

\item{BPPARAM}{instance of a BiocParallel BiocParallelParam subclass}

\item{which_to_use}{vector of elements in ArtifSet (can be package names or numeric indices), defaults to NULL
in which case all elements are examined for events}
}
\description{
scan a slice of an ArtifSet for events produced in R CMD check of source
}
\examples{
cururl = demo_url()
z = setup_artifacts(url=cururl, destbase="test_report")
collect_events(z, event_class="warnings")
collect_events(z, event_class="errors")
}
