% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_artifacts.R
\name{get_process_outcomes}
\alias{get_process_outcomes}
\title{get process outcomes for artifacts}
\usage{
get_process_outcomes(apathset, process)
}
\arguments{
\item{apathset}{instance of artifact_paths}

\item{process}{character(1) one of "buildbin", "buildsrc", "install", "check"}

\item{type}{character(1) one of "EllapsedTime", "Status", "PackageFileSize"}
}
\description{
get process outcomes for artifacts
}
\examples{
eg = select_artifacts(system.file("demo_artifacts/tokay2", package="bbsBuildArtifacts"), 
     c("IRanges", "S4V*", "a4\\\\."))
eg2 = select_artifacts(system.file("demo_artifacts/nebbiolo1", package="bbsBuildArtifacts"), 
     c("IRanges", "S4V*", "a4\\\\."))
get_process_outcomes(eg, process="buildsrc", type="Status")
get_process_outcomes(eg2, "buildsrc", type="EllapsedTime")
get_process_outcomes(eg2, "buildsrc", type="PackageFileSize")
}
