% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_tgz.R
\name{get_report_tgz_cacheid}
\alias{get_report_tgz_cacheid}
\title{get reporting artifacts for a Bioconductor collection (e.g., software, experiment, workflow, ...)}
\usage{
get_report_tgz_cacheid(
  version = "3.14",
  type = "bioc",
  cache = BiocFileCache::BiocFileCache(),
  url = NULL,
  date = Sys.Date()
)
}
\arguments{
\item{version}{character(1) defaults to "3.14"}

\item{type}{character(1) defaults to 'bioc' which implies 'software'; see Note.}

\item{cache}{instance of `BiocFileCache::BiocFileCache()`}

\item{url}{defaults to NULL, if supplied, used to retrieve or cache tgz file}

\item{date}{character(1) defaults to Sys.Date(), "yyyy-mm-dd" format, used to build informative name for caching/cache search}
}
\value{
A gzipped tarball is downloaded, informatively renamed, copied to a cache, and the cache reference is returned.
}
\description{
get reporting artifacts for a Bioconductor collection (e.g., software, experiment, workflow, ...)
}
\note{
Use bbsBuildArtifacts:::valid_types() to see valid values for `type`.  If url is NULL, the
report.tgz content will be cached with name formed with type, version, and date of command invocation.
}
\examples{
cururl = paste0("file://", system.file("test_report_3.14_bioc_20211210/test_report.tgz", 
    package="bbsBuildArtifacts"))
id = get_report_tgz_cacheid(url=cururl)
BiocFileCache::bfcquery(BiocFileCache::BiocFileCache(), cururl)
}
