% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_package_data.R
\name{host_data_by_phase}
\alias{host_data_by_phase}
\title{acquire data on a package's BBS state}
\usage{
host_data_by_phase(bbspd, host, phase)
}
\arguments{
\item{bbspd}{instance of BBS_package_data}

\item{host}{character(1) valid host name for BBS}

\item{phase}{character(1) valid phase}
}
\description{
acquire data on a package's BBS state
}
