% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_package_data.R
\name{make_BBS_package_data}
\alias{make_BBS_package_data}
\title{encapsulate BBS state for a package}
\usage{
make_BBS_package_data(
  afset,
  packagename,
  hosts = c(linux = "nebbiolo2", macos = "machv2", windows = "tokay2")
)
}
\arguments{
\item{afset}{instance of ArtifSet}

\item{packagename}{character(1)}

\item{hosts}{character() valid host names on which BBS was run}
}
\description{
encapsulate BBS state for a package
}
