% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_artifacts.R
\name{make_demo_ArtifSet}
\alias{make_demo_ArtifSet}
\title{simplify production of the test ArtifSet}
\usage{
make_demo_ArtifSet()
}
\description{
simplify production of the test ArtifSet
}
\examples{
make_demo_ArtifSet()
}
