% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_artifacts.R
\name{make_demo_ArtifSet}
\alias{make_demo_ArtifSet}
\title{simplify production of the test ArtifSet}
\usage{
make_demo_ArtifSet(
  cache = BiocFileCache::BiocFileCache(),
  preclean = TRUE,
  demostring = "test_report_3.14_bioc_20211210",
  url = demo_url(),
  destbase = "test_report"
)
}
\arguments{
\item{cache}{BiocFileCache instance to use for checking for available image and retrieving if possible}

\item{preclean}{logical(1) defaults to TRUE, in which case we search `cache` for demostring in cache and remove}

\item{demostring}{character(1) token to query BiocFileCache for in preclean step}

\item{url}{character(1) url for local tgz}

\item{destbase}{character(1) folder name under which all BBS content is held, defaults to `test_report`}
}
\description{
simplify production of the test ArtifSet
}
\examples{
make_demo_ArtifSet()
}
