% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_by_host_data.R
\name{package_by_host_data}
\alias{package_by_host_data}
\title{ingest dcf files for a package}
\usage{
package_by_host_data(
  afpath,
  host = "nebbiolo2",
  summary_types = c("install", "buildsrc", "checksrc", "buildbin"),
  read.dcf.silent = TRUE
)
}
\arguments{
\item{afpath}{an element of an artifact_folder_paths instance}

\item{host}{character(1) host used in BBS}

\item{summary_types}{character() defaults to `c("install", "buildsrc", "checksrc", "buildbin")`}

\item{read.dcf.silent}{logical(1) defaults to TRUE, otherwise bad DCF or build will emit error note}
}
\description{
ingest dcf files for a package
}
\examples{
af = make_demo_ArtifSet()
ho = slot(af, "hostnames")
pbh = package_by_host_data(tail(paths(af),1), host=ho[2])
pbh
package_by_host_data(head(paths(af),1), host=ho[2])
}
