% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_tgz.R
\name{package_by_host_data}
\alias{package_by_host_data}
\title{ingest dcf files for a package}
\usage{
package_by_host_data(
  afpath,
  host = "nebbiolo2",
  summary_types = c("install", "buildsrc", "checksrc"),
  read.dcf.silent = TRUE
)
}
\arguments{
\item{afpath}{an element of an artifact_folder_paths instance}

\item{host}{character(1) host used in BBS}

\item{summary_types}{character() defaults to `c("install", "buildsrc", "checksrc")`}

\item{read.dcf.silent}{logical(1) defaults to TRUE, otherwise bad DCF or build will emit error note}
}
\description{
ingest dcf files for a package
}
\examples{
cururl = demo_url()
af = artifact_folder_paths(url=cururl)
}
