% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_events.R
\name{print.event_collection}
\alias{print.event_collection}
\title{short report of an event collection}
\usage{
\method{print}{event_collection}(x, ...)
}
\description{
short report of an event collection
}
