% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_artifacts.R
\name{setup_artifacts}
\alias{setup_artifacts}
\title{Create ArtifSet instance}
\usage{
setup_artifacts(
  type = "bioc",
  version = "3.14",
  hostnames = hostnames_by_release(version),
  cache = BiocFileCache::BiocFileCache(),
  destination = tempdir(),
  verbose = TRUE,
  extracted = NULL
)
}
\arguments{
\item{type}{character(1) defaults to 'bioc' which implies 'software'; see Note.}

\item{version}{character(1) defaults to "3.14"}

\item{hostnames}{character() vector of host names for which build artifacts are available}

\item{cache}{instance of `BiocFileCache::BiocFileCache()`}

\item{destination}{character(1) path to folder to use for artifacts, defaults to `tempdir()`}

\item{verbose}{logical(1) if TRUE (default) will provide message about tar activity}

\item{extracted}{defaults to NULL, if non-null a character(1) path to folder that holds `report` folder}
}
\description{
Create ArtifSet instance
}
